requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
rasterio>=1.3.0
geopandas>=0.13.0

# Data Handling
pandas>=2.0.0
h5py>=3.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and Analysis
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0